restructured module locations fixed create_entity not returning int id but json fixed errors in step strings
